fix: remove DoiConverter interface
Tests are failing with NoClassDefError when using class with interface
